Update movies.php
These are the text boxes for user input. Once the user clicks submit, the table should populate with what the user entered

However, I'm having a problem with it navigating to the root directory after 'Search' is clicked, rather than running the query and populating the table. I need your help on this one, Michael. Do you know why this is and how it can be fixed?